Resolvendo bug de recebimento de imagem e corpo da requisicao

+++ src/main/java/br/edu/fafic/cz_network/utils/ImageSave.java
@@ -1,5 +1,6 @@
 package br.edu.fafic.cz_network.utils;
 
+import br.edu.fafic.cz_network.model.Usuario;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
@@ -9,10 +10,38 @@ import java.io.OutputStream;
 
 public class ImageSave {
 
-    public static void save(File fileToSave, MultipartFile image) throws IOException {
-        OutputStream os = new FileOutputStream(fileToSave);
-        os.write(image.getBytes());
-        os.close();
+    public void save(MultipartFile image, Usuario usuario, Boolean... postagem) throws IOException {
+
+        String fotoUrl = "C:\\DEVELOP\\JAVA\\cz-network\\src\\main" +
+                "\\java\\br\\edu\\fafic\\cz_network\\imagens\\";
+
+        if (postagem.length > 0) {
+            fotoUrl += "postagem\\" + usuario.getId();
+        } else {
+            fotoUrl += usuario.getId();
+        }
+
+
+        File fileToSave = new File(fotoUrl);
+
+        if (!fileToSave.exists()) {
+            final boolean directoryCreated = fileToSave.mkdirs();
+            if (directoryCreated) {
+                fileToSave = new File(fotoUrl + "\\"
+                        + image.getOriginalFilename());
+
+                OutputStream os = new FileOutputStream(fileToSave);
+                os.write(image.getBytes());
+                os.close();
+            }
+        } else {
+            fileToSave = new File(fotoUrl + "\\"
+                    + image.getOriginalFilename());
+
+            OutputStream os = new FileOutputStream(fileToSave);
+            os.write(image.getBytes());
+            os.close();
+        }
     }
 
 }
